보스 충돌/가서 트리거로 바꾸고 카메라 문제 해결

+++ Assets/Scripts/JiEun/Boss01/BossSkill01.cs
@@ -14,7 +14,7 @@ public class BossSkill01 : MonoBehaviour
     float angle;
     float rotationSpeed;
 
-    int _BossHp;
+    float _BossHp;
     void Start()
     {
